Moved test_commands fixtures to conftest.py
