+++ tests/conftest.py
@@ -3,6 +3,7 @@
 """
 import os
 import shutil
+from pathlib import Path
 
 import pytest
 import sqlalchemy as sa
@@ -94,3 +95,24 @@ def session():
     session.commit()
 
     return session
+
+
+@pytest.fixture
+def package_data():
+    """A dummy set of packages."""
+    data = {'PkgA': ['vers1', 'vers2', 'vers3'],
+            'PkgB': ['versionA', 'VersionB', '0.0.3'],
+            }
+    return data
+
+
+@pytest.fixture
+def package_dir(package_data):
+    """A dummy package directory with packages."""
+    pkg_dir = Path(DATA_DIR) / Path('pkgs')
+    for pkg, versions in package_data.items():
+        for version in versions:
+            path = pkg_dir / Path(pkg) / Path(version)
+            os.makedirs(str(path), exist_ok=True)
+    yield pkg_dir
+    shutil.rmtree(str(pkg_dir), ignore_errors=False)
